backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.25.2
scikit-learn==1.3.2
Werkzeug==2.3.7 